lesson_27_28/project_lesson_27/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-filter==25.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
psycopg==3.3.2
alembic==1.17.2
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
build==1.2.2.post1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==46.0.3
defusedxml==0.7.1
distro==1.9.0
Django==5.2.8
django-admin-ai==0.1.6
django-crispy-forms==2.5
django-filter==25.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
Faker==38.2.0
filelock==3.20.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fsspec==2025.12.0
groq==1.0.0
h11==0.14.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
langdetect==1.0.9
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
oauthlib==3.3.1
openai==2.2.0
opencv-python==4.12.0.88
packaging==24.2
pillow==12.0.0
psycopg==3.3.2
psycopg2-binary==2.9.11
pyclipper==1.4.0
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
python-doctr==0.11.0
python-dotenv==1.2.1
python3-openid==3.2.0
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
requests-oauthlib==2.0.0
scipy==1.16.3
setuptools==75.8.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.3
SQLAlchemy==2.0.44
sqlparse==0.5.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.6.2
Werkzeug==3.1.3
